pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-note"
version = "0.1.0"
description = "Generate data note markdown from BioProject metadata"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "biopython",
  "jinja2",
  "num2words",
  "pandas",
  "Pillow",
  "python-dateutil",
  "PyYAML",
  "requests",
  "tenacity",
]

[project.scripts]
data-note = "data_note.cli:main"

[tool.setuptools.packages.find]
include = ["data_note*"]
